+++ src/chat21-core/providers/firebase/firebase-typing.service.ts
@@ -56,8 +56,8 @@ export class FirebaseTypingService extends TypingService {
     ref.on('child_changed', (childSnapshot) => {
       const precence: TypingModel = childSnapshot.val();
       //TODO-GAB: supportata solo da chat ionic 5
-      // this.BSIsTyping.next({uid: idConversation, uidUserTypingNow: childSnapshot.val().user.uid, nameUserTypingNow: childSnapshot.val().user.name});
-      this.BSIsTyping.next({uid: idConversation, uidUserTypingNow: precence.uid, nameUserTypingNow: precence.name});
+      //this.BSIsTyping.next({uid: idConversation, uidUserTypingNow: precence.uid, nameUserTypingNow: precence.name});
+      this.BSIsTyping.next({uid: idConversation, uidUserTypingNow: childSnapshot.key, nameUserTypingNow: precence.name});
     });
   }
 
@@ -79,5 +79,6 @@ export class FirebaseTypingService extends TypingService {
       });
     }, 500);
   }
+
 
 }